Change the page title to prioritize page content
When many tab's are open (which is necessary for DRF's documentation), it becomes impossible to determine which tab contains which pieces of documentation.  That they are all related is obvious by the use of a common icon, just not the specific page each tab is loaded to.

This change inverts the order of the title to maintain as much useful context as possible on the tabbar.